README update,text.h added,cpu is "smarter"
Added an another level for cpu. They don't choose everything randomly
in this level.
Added text.h,moved most of the strings to here. May move more to here.
Updated editor,you can open it directly or open it in the menu of the
game.
Updated READEME.added some tutorial to it.